chore(add-use-spacing): bumping version and moving tests
