--- src/@types/styled-components.d.ts
@@ -1,7 +1,0 @@
-import 'styled-components';
-
-declare module 'styled-components' {
-  export interface DefaultTheme {
-    white: string;
-  }
-}
