+++ src/index.ts
@@ -1,1 +1,9 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  export interface DefaultTheme {
+    white: string;
+  }
+}
+
 export * from './web';
