+++ src/shared/__tests__/toOnlyNumbers.spec.ts
@@ -1,4 +1,4 @@
-import { toOnlyNumbers } from '../shared/toOnlyNumbers';
+import { toOnlyNumbers } from '../toOnlyNumbers';
 
 describe('toOnlyNumbers tests', () => {
   it('should format to string empty properly', () => {
